+++ src/views/dashboard/components/Pagination.vue
@@ -1,0 +1,8 @@
+<script setup lang="ts"></script>
+
+
+<template>
+
+</template>
+
+<style scoped lang="less"></style>
